Updated page switching and made the scraper slightly faster
For some reason, an update before had made page switching crash the scraper because the sources intended to be retrieved on the next page do not wait to be loaded. Added a wait.
Also, most importantly, there was an issue with documents being scraped where only the first few documents would be scraped in addition to the fact that more than usual number of documents would be scraped.