Update chrome link and requirements
Updated chrome link to download latest chrome browser as lately the GUI has crashed from its inability to find a chrome browser type. This may indicate that a chrome browser is being newly downloaded each time and that this is a source of failure when the chrome browser is no longer accessible for one reason or another.  For that reason, a more future proof oriented approach may be warranted where the browser is predownloaded and installed. At this time, I do not have enough time to seek out this solution as previous cursory attempts have failed. Therefore, if you are reading this, be aware that this current GUI may become inoperable as time goes on especially if eHRAF changes their website.

Additionally, redid the requirements.txt to add 'pyinstaller' which originally was not there. Additionally this updated requirements.txt has updated packages