eHRAF change hot fix
I believe eHRAF's website has made a change to the way they display files making the scraper bricked. This hot fix disables one check to make sure the scraper is working and alters a few code segments to make the code work again. Please be aware this has not be checked for issues that this may have caused! Stability will be added later